refactor: move rate limiter to separate file + add cleanup
Per review feedback:
- Moved rate limiting logic to src/server/lib/rate-limit.ts
- Added hourly cleanup scheduler to prevent loginAttempts from growing
  infinitely - removes expired records every hour
